+++ src/Curl/CaseInsensitiveArray.php
@@ -42,12 +42,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @param  mixed[]              $initial (optional) Existing array to convert.
      * @return CaseInsensitiveArray
      */
-    // TODO: Use a nullable type declaration when supported versions >= PHP 7.1.
-    //   Trying to use the nullable type declaration on PHP 7.0:
-    //     public function __construct(?array $initial = null)
-    //   results in:
-    //     ParseError: syntax error, unexpected '?', expecting variable (T_VARIABLE)
-    public function __construct($initial = null)
+    public function __construct(?array $initial = null)
     {
         if ($initial !== null) {
             foreach ($initial as $key => $value) {
@@ -68,6 +63,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @return void
      * @see https://secure.php.net/manual/en/arrayaccess.offsetset.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function offsetSet($offset, $value)
     {
@@ -90,10 +86,11 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @return bool   If the offset exists.
      * @see https://secure.php.net/manual/en/arrayaccess.offsetexists.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function offsetExists($offset)
     {
-        return (bool) array_key_exists(strtolower($offset), $this->data);
+        return array_key_exists(strtolower($offset), $this->data);
     }
 
     /**
@@ -106,6 +103,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @return void
      * @see https://secure.php.net/manual/en/arrayaccess.offsetunset.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function offsetUnset($offset)
     {
@@ -124,6 +122,7 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @return mixed  The data stored at the offset.
      * @see https://secure.php.net/manual/en/arrayaccess.offsetget.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function offsetGet($offset)
     {
@@ -134,23 +133,23 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     /**
      * Count
      *
-     * @param void
      * @return int The number of elements stored in the array.
      * @see https://secure.php.net/manual/en/countable.count.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function count()
     {
-        return (int) count($this->data);
+        return count($this->data);
     }
 
     /**
      * Current
      *
-     * @param void
      * @return mixed Data at the current position.
      * @see https://secure.php.net/manual/en/iterator.current.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function current()
     {
@@ -160,10 +159,10 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     /**
      * Next
      *
-     * @param void
      * @return void
      * @see https://secure.php.net/manual/en/iterator.next.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function next()
     {
@@ -173,15 +172,19 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
     /**
      * Key
      *
-     * @param void
      * @return mixed Case-sensitive key at current position.
      * @see https://secure.php.net/manual/en/iterator.key.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function key()
     {
         $key = key($this->data);
-        return $this->keys[$key] ?? $key;
+        if ($key === null) {
+            return null;
+        } else {
+            return $this->keys[$key] ?? $key;
+        }
     }
 
     /**
@@ -190,19 +193,20 @@ class CaseInsensitiveArray implements \ArrayAccess, \Countable, \Iterator
      * @return bool If the current position is valid.
      * @see https://secure.php.net/manual/en/iterator.valid.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function valid()
     {
-        return (bool) (key($this->data) !== null);
+        return (key($this->data) !== null);
     }
 
     /**
      * Rewind
      *
-     * @param void
      * @return void
      * @see https://secure.php.net/manual/en/iterator.rewind.php
      */
+    #[\Override]
     #[\ReturnTypeWillChange]
     public function rewind()
     {