+++ tests/Helper.php
@@ -8,8 +8,8 @@ use Curl\Curl;
 
 class Test
 {
-    const TEST_URL = 'http://127.0.0.1:8000/';
-    const ERROR_URL = 'http://1.2.3.4/';
+    public const TEST_URL = 'http://127.0.0.1:8000/';
+    public const ERROR_URL = 'http://1.2.3.4/';
 
     public $message;
 
@@ -123,7 +123,6 @@ function mime_type($file_path)
 {
     $finfo = finfo_open(FILEINFO_MIME_TYPE);
     $mime_type = finfo_file($finfo, $file_path);
-    finfo_close($finfo);
     return $mime_type;
 }