+++ tests/PHPCurlClass/PHPCurlClassTest.php
@@ -3187,23 +3187,13 @@ class PHPCurlClassTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * @requires PHPUnit < 10
-     * @expectedException \PHPUnit\Framework\Error\Warning
-     */
-    public function testRequiredOptionCurlOptReturnTransferEmitsWarning()
-    {
-        $this->expectWarning(\PHPUnit\Framework\Error\Warning::class);
-
-        $curl = new Curl();
-        $curl->setOpt(CURLOPT_RETURNTRANSFER, false);
-    }
-
-    /**
      * @requires PHPUnit >= 10
      */
+    #[RequiresPhpunit('>= 10')]
     public function testRequiredOptionCurlOptReturnTransferEmitsWarningPHPUnit10Plus()
     {
         set_error_handler(static function (int $errno, string $errstr): never {
+            restore_error_handler();
             throw new \Exception($errstr, $errno);
         }, E_USER_WARNING);
 
@@ -3211,8 +3201,6 @@ class PHPCurlClassTest extends \PHPUnit\Framework\TestCase
 
         $curl = new Curl();
         $curl->setOpt(CURLOPT_RETURNTRANSFER, false);
-
-        restore_error_handler();
     }
 
     public function testRequestMethodSuccessiveGetRequests()
@@ -3401,12 +3389,6 @@ class PHPCurlClassTest extends \PHPUnit\Framework\TestCase
 
     public function testMemoryLeak()
     {
-        // Skip memory leak test failing for PHP 7.
-        // "Failed asserting that 8192 is less than 1000."
-        if (getenv('CI_PHP_VERSION') === '7.0') {
-            $this->markTestSkipped();
-        }
-
         ob_start();
         echo '[';
         for ($i = 0; $i < 10; $i++) {
@@ -3972,7 +3954,7 @@ class PHPCurlClassTest extends \PHPUnit\Framework\TestCase
 
         $curl->expects($this->once())
              ->method('getRawResponse')
-             ->will($this->returnValue('[]'));
+             ->willReturn('[]');
 
         $this->assertEquals('[]', $curl->getRawResponse());
     }