+++ tests/run_phpunit.sh
@@ -27,9 +27,9 @@ replace_assertMatchesRegularExpression() {
     find='->assertMatchesRegularExpression('
     replace='->assertRegExp('
     if sed v < /dev/null 2> /dev/null; then
-        sed -i"" -e"s/${find}/${replace}/" "./PHPCurlClass/PHP"*
+        sed -i"" -e "s/${find}/${replace}/" "./PHPCurlClass/PHP"*
     else
-        sed -i "" -e"s/${find}/${replace}/" "./PHPCurlClass/PHP"*
+        sed -i "" -e "s/${find}/${replace}/" "./PHPCurlClass/PHP"*
     fi
 }
 
@@ -57,33 +57,16 @@ phpunit_v10_shim() {
     remove_expectWarning
 }
 
-php_v7_0_shim() {
-    # -protected function setUp(): void
-    # +protected function setUp()
-    find='protected function setUp(): void'
-    replace='protected function setUp()'
-    if sed v < /dev/null 2> /dev/null; then
-        sed -i"" -e"s/${find}/${replace}/" "./PHPCurlClass/PHP"*
-    else
-        sed -i "" -e"s/${find}/${replace}/" "./PHPCurlClass/PHP"*
-    fi
+phpunit_v11_shim() {
+    :;
 }
 
-set -x
-
-composer self-update
-composer install --prefer-source --no-interaction
-
-# Use composer's phpunit and phpcs by adding composer bin directory to the path environment variable.
-export PATH="${PWD}/vendor/bin:${PATH}"
+phpunit_v12_shim() {
+    remove_expectWarning
+}
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 cd "${SCRIPT_DIR}"
-
-echo "CI_PHP_VERSION: ${CI_PHP_VERSION}"
-echo "CI_PHP_FUTURE_RELEASE: ${CI_PHP_FUTURE_RELEASE}"
-php -r "var_dump(phpversion());"
-php -r "var_dump(curl_version());"
 
 # Let test server know we should allow testing.
 export PHP_CURL_CLASS_TEST_MODE_ENABLED="yes"
@@ -91,18 +74,14 @@ export PHP_CURL_CLASS_TEST_MODE_ENABLED="yes"
 # Start test servers. Run servers on different ports to allow simultaneous
 # requests without blocking.
 server_count=7
-declare -A pids
-for i in $(seq 0 $(("${server_count}" - 1))); do
+pids=()
+for i in $(seq 0 "$(echo "${server_count} - 1" | bc)"); do
     port=8000
     (( port += $i ))
 
     php -S "127.0.0.1:${port}" server.php &> /dev/null &
-    pids["${i}"]="${!}"
+    pids+=("${!}")
 done
-
-errors=()
-
-source "check_syntax.sh"
 
 # Determine which phpunit to use.
 if [[ -f "../vendor/bin/phpunit" ]]; then
@@ -130,10 +109,12 @@ elif [[ "${phpunit_version}" == "9."* ]]; then
 elif [[ "${phpunit_version}" == "10."* ]]; then
     phpunit_v10_shim
     phpunit_args=" --display-incomplete --display-skipped --display-deprecations --display-errors --display-notices --display-warnings --fail-on-risky ${extra_args}"
-fi
-
-if [[ "${CI_PHP_VERSION}" == "7.0" ]]; then
-    php_v7_0_shim
+elif [[ "${phpunit_version}" == "11."* ]]; then
+    phpunit_v11_shim
+    phpunit_args=" --display-incomplete --display-skipped --display-deprecations --display-errors --display-notices --display-warnings --fail-on-risky ${extra_args}"
+elif [[ "${phpunit_version}" == "12."* ]]; then
+    phpunit_v12_shim
+    phpunit_args=" --display-incomplete --display-skipped --display-deprecations --display-errors --display-notices --display-warnings --fail-on-risky ${extra_args}"
 fi
 
 # Run tests.
@@ -142,33 +123,11 @@ fi
     --configuration "phpunit.xml" \
     ${phpunit_args}
 if [[ "${?}" -ne 0 ]]; then
-    echo "Error: phpunit command failed"
+    echo "❌ Error: phpunit command failed"
     errors+=("phpunit command failed")
-fi
-
-source "check_coding_standards.sh"
-
-set +x
-
-error_count="${#errors[@]}"
-if [[ "${error_count}" -ge 1 ]]; then
-    echo -e "\nErrors found: ${error_count}"
-
-    iter=0
-    for value in "${errors[@]}"; do
-        ((iter++))
-        echo -e "\nError ${iter} of ${error_count}:"
-        echo "${value}" | perl -pe 's/^(.*)$/\t\1/'
-    done
 fi
 
 # Stop test servers.
 for pid in "${pids[@]}"; do
-  kill "${pid}" &> /dev/null &
+    kill "${pid}" &> /dev/null &
 done
-
-if [[ "${CI_PHP_FUTURE_RELEASE}" != "true" ]]; then
-    exit "${#errors[@]}"
-elif [[ "${#errors[@]}" -ne 0 ]]; then
-    echo "One or more tests failed, but allowed as the CI_PHP_FUTURE_RELEASE flag is on for PHP version ${CI_PHP_VERSION}."
-fi